requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Data handling
networkx>=3.1
ogb>=1.3.6

# Visualization and plotting
matplotlib>=3.7.0
plotly>=5.15.0
pyvis>=0.3.2

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Evaluation and metrics
torchmetrics>=1.0.0

# Optional: logging and experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Development and formatting
black>=23.0.0
ruff>=0.0.280
pre-commit>=3.3.0

# Demo and serving
streamlit>=1.25.0
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
